organize files in app/src/utils directory
