+++ app/src/utils/exercises/MuscleUp.js
@@ -1,4 +1,4 @@
-import { genCheck } from './GenFeedback';
+import { genCheck } from '../GenFeedback';
 
 const muscleUpInfo = {
     states: {
