+++ app/src/utils/exercises/PullUp.js
@@ -1,4 +1,4 @@
-import { genCheck } from './GenFeedback';
+import { genCheck } from '../GenFeedback';
 
 const pullUpInfo = {
     states: {
