+++ app/src/utils/helpers/ExerciseSettings.js
@@ -1,5 +1,5 @@
 import { doc, setDoc, getDoc } from "firebase/firestore";
-import { db } from "../firebaseConfig";
+import { db } from "../../firebaseConfig";
 
 /**
  * Stores exercise settings for a user in the Firestore database.
